Amesbury Testate Amoebae dataset

Dataset from Amesbury (2016) in Development of a new pan-European testate amoeba transfer function for reconstructing peatland palaeohydrology Report

Usage

Amesbury

Format

Amesbury

This dataset contains the data from Amesbury (2016). In essence, it's a Testate amoebae community table (45 broad TA taxa and 1103 samples)

spe.df

A species x sites dataframe with stites as rows and species in column

WTD

A vector od Water table depth associated with each samples

coords

a dataframe containing the coordinates of each sample

Source

doi:10.1016/j.quascirev.2016.09.024

References

Matthew J. Amesbury, Graeme T. Swindles, Anatoly Bobrov, Dan J. Charman, Joseph Holden, Mariusz Lamentowicz, Gunnar Mallon, Yuri Mazei, Edward A.D. Mitchell, Richard J. Payne, Thomas P. Roland, T. Edward Turner, Barry G. Warner, Development of a new pan-European testate amoeba transfer function for reconstructing peatland palaeohydrology. Quaternary Science Reviews, vol. 152, 2016, pages 132-151. doi:10.1016/j.quascirev.2016.09.024 .